# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Position {
  x?: number | null;
  y?: number | null;
  z?: number | null;
}

export interface AttackMode {
  enabled: boolean;
  targetPlayer?: string | null;
  attackMobs: boolean;
}

export interface AntiAfk {
  enabled: boolean;
}

export interface FollowMode {
  enabled: boolean;
  targetPlayer?: string | null;
}

export interface AutoDrop {
  enabled: boolean;
}

export interface BotStatus {
  connected: boolean;
  state: string;
  host?: string | null;
  port?: number | null;
  username?: string | null;
  version?: string | null;
  health?: number | null;
  food?: number | null;
  position?: Position | null;
  attackMode: AttackMode;
  antiAfk: AntiAfk;
  followMode: FollowMode;
  autoDrop: AutoDrop;
  ping?: number | null;
}

export interface ConnectBotRequest {
  host: string;
  port?: number;
  username: string;
  version: string;
  password?: string | null;
}

export interface AttackModeRequest {
  enabled: boolean;
  targetPlayer?: string | null;
  attackMobs: boolean;
}

export interface AntiAfkRequest {
  enabled: boolean;
}

export interface FollowModeRequest {
  enabled: boolean;
  targetPlayer?: string | null;
}

export interface AutoDropRequest {
  enabled: boolean;
}

export interface SendChatRequest {
  message: string;
}

export type ConsoleLogType = typeof ConsoleLogType[keyof typeof ConsoleLogType];


export const ConsoleLogType = {
  info: 'info',
  error: 'error',
  chat: 'chat',
  system: 'system',
} as const;

export interface ConsoleLog {
  id: string;
  timestamp: string;
  type: ConsoleLogType;
  message: string;
}

export interface MemoryEntry {
  server: string;
  registered: boolean;
  password?: string | null;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetConsoleParams = {
limit?: number;
};