# lib/api-zod/src/generated/types/botStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AntiAfk } from './antiAfk';
import type { AttackMode } from './attackMode';
import type { AutoDrop } from './autoDrop';
import type { FollowMode } from './followMode';
import type { Position } from './position';

export interface BotStatus {
  connected: boolean;
  state: string;
  host?: string | null;
  port?: number | null;
  username?: string | null;
  version?: string | null;
  health?: number | null;
  food?: number | null;
  position?: Position | null;
  attackMode: AttackMode;
  antiAfk: AntiAfk;
  followMode: FollowMode;
  autoDrop: AutoDrop;
  ping?: number | null;
}